fix: filters, moved item to separate file, reseting nested items
