+++ src/components/frame/Extensions/pages/PermissionsPage/partials/PermissionsPageRow.tsx
@@ -78,8 +78,9 @@ const PermissionsPageRow: FC<PermissionPageRowProps> = ({
         <EmptyContent
           title={formatText({ id: 'permissionsPage.empty.title' })}
           description={formatText({ id: 'permissionsPage.empty.description' })}
-          className="px-4 py-[1.375rem]"
+          className="pt-10 pb-9 px-6 mt-6"
           icon={Binoculars}
+          withBorder
         />
       )}
     </div>
